requirements.txt
alembic==1.12.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
Werkzeug==3.0.1
